Moved filtering to a custom hook, which I applied in the App component
